requirements.txt
flask
opencv-python
numpy
scipy
imutils
torch
torchvision
pillow
